feat: 完成方法组合
